# docs/swagger.yaml
definitions:
...
  internal_controllers.DocumentResponse:
    properties:
      content:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
  description: Сервис для загрузки документов, подсчёта TF‑IDF и управления коллекциями.
  title: LestaStartTest API
  version: 2.2.1
paths:
...
  /api/documents:
    get:
      description: Возвращает все документы текущего пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: '{"documents":[]DocumentResponse}'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Database error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список документов
      tags:
      - Документы
  /api/documents/{id}:
    delete:
      description: Удаляет документ, все связи и физический файл.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"Document deleted"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Document not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete document
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление документа
      tags:
      - Документы
    get:
      description: Возвращает имя и содержимое документа по его ID.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controllers.DocumentResponse'
        "404":
          description: Document not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение документа
      tags:
      - Документы
  /api/documents/{id}/huffman:
    get:
      description: Возвращает закодированное представление содержимого документа.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"document_id":int,"huffman_encoded":string}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid document ID or content too large
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Document not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Encoding failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Кодирование документа алгоритмом Хаффмана
      tags:
      - Документы
  /api/documents/{id}/statistics:
    get:
      description: Рассчитывает TF‑IDF слова внутри всех коллекций, где есть этот
        документ.
      parameters:
      - description: ID документа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"document_id":int,"statistics":map[string]object}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Document is not in any collection
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Document not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to find collections
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: TF‑IDF статистика документа
      tags:
      - Документы
  /api/documents/upload:
    post:
      consumes:
      - multipart/form-data
      description: Загружает один или несколько файлов, обрабатывает и сохраняет их.
      parameters:
      - collectionFormat: multi
        description: Файлы для загрузки
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":string,"data":UploadResult}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error getting files or no files uploaded
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: '{"errors":[]string}'
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Загрузка файлов
      tags:
      - Документы
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"